moving guard check and reset from scenario.py to simulator and verifier
