+++ verse/analysis/verifier.py
@@ -1,17 +1,55 @@
+from collections import defaultdict
 import copy, itertools, functools, pprint, ray
 from typing import Dict, List, Optional, Tuple
 import numpy as np
+import warnings 
+import ast
+
 # from verse.agents.base_agent import BaseAgent
 from verse.analysis.analysis_tree import AnalysisTreeNode, AnalysisTree
 from verse.analysis.dryvr import calc_bloated_tube, SIMTRACENUM
 from verse.analysis.mixmonotone import calculate_bloated_tube_mixmono_cont, calculate_bloated_tube_mixmono_disc
 from verse.analysis.incremental import ReachTubeCache, TubeCache, convert_reach_trans, to_simulate, combine_all
+from verse.analysis.incremental import CachedRTTrans, CachedSegment, combine_all, reach_trans_suit, sim_trans_suit
 from verse.analysis.utils import dedup
 from verse.map.lane_map import LaneMap
-from verse.parser.parser import find
+from verse.parser.parser import find,ModePath
+from verse.agents.base_agent import BaseAgent
+from verse.automaton import GuardExpressionAst, ResetExpression
+
 pp = functools.partial(pprint.pprint, compact=True, width=130)
 
 import time
+
+PathDiffs = List[Tuple[BaseAgent,ModePath]]
+EGO, OTHERS = "ego", "others"
+
+def combine_rect_v1(trace):
+    trace = np.array(trace)
+    num_step, num_dim = trace.shape
+    assert num_step % 2 == 0
+    combined_trace = np.ndarray(shape=(2, num_dim))
+    combined_trace[0] = np.min(trace[::2], 0)
+    combined_trace[1] = np.max(trace[1::2], 0)
+    return combined_trace.tolist()
+
+def combine_rect_v2(trace: np.ndarray, rect_len):
+    trace = np.array(trace)
+    num_step, num_dim = trace.shape
+    assert num_step % 2 == 0
+    lower=trace[::2]
+    upper=trace[1::2]
+    total_rect_num = num_step / 2
+    combined_rect_num = int(np.ceil(total_rect_num/rect_len))
+    combined_trace = np.ndarray(shape=(2*combined_rect_num, num_dim))
+    for i in range(combined_rect_num-1):
+        combined_trace[2*i] = np.min(lower[i*rect_len: (i+1)*rect_len], 0)[1:]
+        combined_trace[2*i+1] = np.max(upper[i*rect_len: (i+1)*rect_len], 0)[1:]
+    i = combined_rect_num-1
+    combined_trace[2*i] = np.min(lower[i*rect_len:], 0)
+    combined_trace[2*i+1] = np.max(upper[i*rect_len:], 0)
+    return combined_trace.tolist()
+
 
 class Verifier:
     def __init__(self, config):
@@ -110,7 +148,7 @@ class Verifier:
         reachability_method,
         run_num: int,
         past_runs: List[AnalysisTree],
-        transition_graph,
+        sensor,
         params = {},
     ):
         if (max_height == None):
@@ -203,7 +241,7 @@ class Verifier:
                 # pp(("to sim", new_cache.keys(), len(paths_to_sim)))
 
         # Get all possible transitions to next mode
-        asserts, all_possible_transitions = transition_graph.get_transition_verify_opt(new_cache, paths_to_sim, node)
+        asserts, all_possible_transitions = self.get_transition_verify_opt(new_cache, paths_to_sim, node, lane_map, sensor)
         node.assert_hits = asserts
 
         if asserts != None:
@@ -295,7 +333,7 @@ class Verifier:
         static_list: List[str],
         uncertain_param_list: List[float],
         agent_list,
-        transition_graph,
+        sensor,
         time_horizon,
         time_step,
         max_height,
@@ -347,7 +385,7 @@ class Verifier:
                     continue
                 # For trace not already verified
                 result_refs.append(self.compute_full_reachtube_step.remote
-                                (self, node, remain_time, time_step, max_height, lane_map, init_seg_length, reachability_method,run_num, past_runs, transition_graph, params))
+                                (self, node, remain_time, time_step, max_height, lane_map, init_seg_length, reachability_method,run_num, past_runs, sensor, params))
                 if len(result_refs) >= self.config.parallel_ver_ahead:
                     wait = True
             elif len(result_refs) > 0:
@@ -384,7 +422,7 @@ class Verifier:
         static_list: List[str],
         uncertain_param_list: List[float],
         agent_list,
-        transition_graph,
+        sensor,
         time_horizon,
         time_step,
         max_height,
@@ -528,7 +566,7 @@ class Verifier:
             # asserts_orig, all_possible_transitions_orig = transition_graph.get_transition_verify(new_cache, paths_to_sim, node)
             # time_transition_orig += time.perf_counter() - start
             start = time.perf_counter()
-            asserts, all_possible_transitions = transition_graph.get_transition_verify_opt(new_cache, paths_to_sim, node)
+            asserts, all_possible_transitions = self.get_transition_verify_opt(new_cache, paths_to_sim, node, lane_map, sensor)
             time_transition += time.perf_counter() - start
             # assert asserts == asserts_orig and all_possible_transitions_orig == all_possible_transitions
             # pp(("transitions:", [(t[0], t[2]) for t in all_possible_transitions]))
@@ -619,6 +657,543 @@ class Verifier:
         self.num_transitions = num_transitions
         print('reachtube time:', time_reachtube, 'transition time:', time_transition_orig, 'transition opt time:', time_transition)
         return self.reachtube_tree
+
+    def apply_cont_var_updater(self, cont_var_dict, updater):
+        for variable in updater:
+            for unrolled_variable, unrolled_variable_index in updater[variable]:
+                cont_var_dict[unrolled_variable] = cont_var_dict[variable][unrolled_variable_index]
+
+    def get_transition_verify(self, cache: Dict[str, CachedRTTrans], paths: PathDiffs, node: AnalysisTreeNode, track_map, sensor) -> Tuple[Optional[Dict[str, List[str]]], Optional[Dict[str, List[Tuple[str, List[str], List[float]]]]]]:
+        # For each agent
+        agent_guard_dict = defaultdict(list)
+        cached_guards = defaultdict(list)
+        min_trans_ind = None
+        cached_trans = defaultdict(list)
+        agent_dict = node.agent
+
+        if not cache:
+            paths = [(agent, p) for agent in node.agent.values() for p in agent.decision_logic.paths]
+        else:
+
+            # _transitions = [trans.transition for seg in cache.values() for trans in seg.transitions]
+            _transitions = [(aid, trans) for aid, seg in cache.items() for trans in seg.transitions if reach_trans_suit(trans.inits, node.init)]
+            # pp(("cached trans", len(_transitions)))
+            if len(_transitions) > 0:
+                min_trans_ind = min([t.transition for _, t in _transitions])
+                # TODO: check for asserts
+                cached_trans = [(aid, tran.mode, tran.dest, tran.reset, tran.reset_idx, tran.paths) for aid, tran in dedup(_transitions, lambda p: (p[0], p[1].mode, p[1].dest)) if tran.transition == min_trans_ind]
+                if len(paths) == 0:
+                    # print(red("full cache"))
+                    return None, cached_trans
+
+                path_transitions = defaultdict(int)
+                for seg in cache.values():
+                    for tran in seg.transitions:
+                        for p in tran.paths:
+                            path_transitions[p.cond] = max(path_transitions[p.cond], tran.transition)
+                for agent_id, segment in cache.items():
+                    agent = node.agent[agent_id]
+                    if len(agent.decision_logic.args) == 0:
+                        continue
+                    state_dict = {aid: (node.trace[aid][0], node.mode[aid], node.static[aid]) for aid in node.agent}
+
+                    agent_paths = dedup([p for tran in segment.transitions for p in tran.paths], lambda i: (i.var, i.cond, i.val))
+                    for path in agent_paths:
+                        cont_var_dict_template, discrete_variable_dict, length_dict = sensor.sense(
+                            agent, state_dict, track_map)
+                        reset = (path.var, path.val_veri)
+                        guard_expression = GuardExpressionAst([path.cond_veri])
+
+                        cont_var_updater = guard_expression.parse_any_all_new(
+                            cont_var_dict_template, discrete_variable_dict, length_dict)
+                        self.apply_cont_var_updater(
+                            cont_var_dict_template, cont_var_updater)
+                        guard_can_satisfied = guard_expression.evaluate_guard_disc(
+                            agent, discrete_variable_dict, cont_var_dict_template, track_map)
+                        if not guard_can_satisfied:
+                            continue
+                        cached_guards[agent_id].append((path, guard_expression, cont_var_updater, copy.deepcopy(discrete_variable_dict), reset, path_transitions[path.cond]))
+
+        # for aid, trace in node.trace.items():
+        #     if len(trace) < 2:
+        #         pp(("weird state", aid, trace))
+        for agent, path in paths:
+            if len(agent.decision_logic.args) == 0:
+                continue
+            agent_id = agent.id
+            state_dict = {aid: (node.trace[aid][0:2], node.mode[aid], node.static[aid]) for aid in node.agent}
+            cont_var_dict_template, discrete_variable_dict, length_dict = sensor.sense(
+                agent, state_dict, track_map)
+            # TODO-PARSER: Get equivalent for this function
+            # Construct the guard expression
+            guard_expression = GuardExpressionAst([path.cond_veri])
+
+            cont_var_updater = guard_expression.parse_any_all_new(
+                cont_var_dict_template, discrete_variable_dict, length_dict)
+            self.apply_cont_var_updater(
+                cont_var_dict_template, cont_var_updater)
+            guard_can_satisfied = guard_expression.evaluate_guard_disc(
+                agent, discrete_variable_dict, cont_var_dict_template, track_map)
+            if not guard_can_satisfied:
+                continue
+            agent_guard_dict[agent_id].append(
+                (guard_expression, cont_var_updater, copy.deepcopy(discrete_variable_dict), path))
+
+        trace_length = int(min(len(v) for v in node.trace.values()) // 2)
+        # pp(("trace len", trace_length, {a: len(t) for a, t in node.trace.items()}))
+        guard_hits = []
+        guard_hit = False
+        for idx in range(trace_length):
+            if min_trans_ind != None and idx >= min_trans_ind:
+                return None, cached_trans
+            any_contained = False
+            hits = []
+            state_dict = {aid: (node.trace[aid][idx*2:idx*2+2], node.mode[aid], node.static[aid]) for aid in node.agent}
+
+            asserts = defaultdict(list)
+            for agent_id in agent_dict.keys():
+                agent: BaseAgent = agent_dict[agent_id]
+                if len(agent.decision_logic.args) == 0:
+                    continue
+                agent_state, agent_mode, agent_static = state_dict[agent_id]
+                # if np.array(agent_state).ndim != 2:
+                #     pp(("weird state", agent_id, agent_state))
+                agent_state = agent_state[1:]
+                cont_vars, disc_vars, len_dict = sensor.sense(agent, state_dict, track_map)
+                resets = defaultdict(list)
+                # Check safety conditions
+                for i, a in enumerate(agent.decision_logic.asserts_veri):
+                    pre_expr = a.pre
+
+                    def eval_expr(expr):
+                        ge = GuardExpressionAst([copy.deepcopy(expr)])
+                        cont_var_updater = ge.parse_any_all_new(cont_vars, disc_vars, len_dict)
+                        self.apply_cont_var_updater(cont_vars, cont_var_updater)
+                        sat = ge.evaluate_guard_disc(agent, disc_vars, cont_vars, track_map)
+                        if sat:
+                            sat = ge.evaluate_guard_hybrid(agent, disc_vars, cont_vars, track_map)
+                            if sat:
+                                sat, contained = ge.evaluate_guard_cont(agent, cont_vars, track_map)
+                                sat = sat and contained
+                        return sat
+                    if eval_expr(pre_expr):
+                        if not eval_expr(a.cond):
+                            label = a.label if a.label != None else f"<assert {i}>"
+                            print(f"assert hit for {agent_id}: \"{label}\"")
+                            print(idx)
+                            asserts[agent_id].append(label)
+                if agent_id in asserts:
+                    continue
+                if agent_id not in agent_guard_dict:
+                    continue
+
+                unchecked_cache_guards = [g[:-1] for g in cached_guards[agent_id] if g[-1] < idx]     # FIXME: off by 1?
+                for guard_expression, continuous_variable_updater, discrete_variable_dict, path in agent_guard_dict[agent_id] + unchecked_cache_guards:
+                    assert isinstance(path, ModePath)
+                    new_cont_var_dict = copy.deepcopy(cont_vars)
+                    one_step_guard: GuardExpressionAst = copy.deepcopy(guard_expression)
+
+                    self.apply_cont_var_updater(new_cont_var_dict, continuous_variable_updater)
+                    guard_can_satisfied = one_step_guard.evaluate_guard_hybrid(
+                        agent, discrete_variable_dict, new_cont_var_dict, track_map)
+                    if not guard_can_satisfied:
+                        continue
+                    guard_satisfied, is_contained = one_step_guard.evaluate_guard_cont(
+                        agent, new_cont_var_dict, track_map)
+                    any_contained = any_contained or is_contained
+                    # TODO: Can we also store the cont and disc var dict so we don't have to call sensor again?
+                    if guard_satisfied:
+                        reset_expr = ResetExpression((path.var, path.val_veri))
+                        resets[reset_expr.var].append(
+                            (reset_expr, discrete_variable_dict,
+                             new_cont_var_dict, guard_expression.guard_idx, path)
+                        )
+                # Perform combination over all possible resets to generate all possible real resets
+                combined_reset_list = list(itertools.product(*resets.values()))
+                if len(combined_reset_list) == 1 and combined_reset_list[0] == ():
+                    continue
+                for i in range(len(combined_reset_list)):
+                    # Compute reset_idx
+                    reset_idx = []
+                    for reset_info in combined_reset_list[i]:
+                        reset_idx.append(reset_info[3])
+                    # a list of reset expression
+                    hits.append((agent_id, tuple(reset_idx), combined_reset_list[i]))
+            if len(asserts) > 0:
+                return (asserts, idx), None
+            if hits != []:
+                guard_hits.append((hits, state_dict, idx))
+                guard_hit = True
+            elif guard_hit:
+                break
+            if any_contained:
+                break
+
+        reset_dict = {}  # defaultdict(lambda: defaultdict(list))
+        for hits, all_agent_state, hit_idx in guard_hits:
+            for agent_id, reset_idx, reset_list in hits:
+                # TODO: Need to change this function to handle the new reset expression and then I am done
+                dest_list, reset_rect = self.apply_reset(node.agent[agent_id], reset_list, all_agent_state, track_map)
+                # pp(("dests", dest_list, *[astunparser.unparse(reset[-1].val_veri) for reset in reset_list]))
+                if agent_id not in reset_dict:
+                    reset_dict[agent_id] = {}
+                if not dest_list:
+                    warnings.warn(
+                        f"Guard hit for mode {node.mode[agent_id]} for agent {agent_id} without available next mode")
+                    dest_list.append(None)
+                if reset_idx not in reset_dict[agent_id]:
+                    reset_dict[agent_id][reset_idx] = {}
+                for dest in dest_list:
+                    if dest not in reset_dict[agent_id][reset_idx]:
+                        reset_dict[agent_id][reset_idx][dest] = []
+                    reset_dict[agent_id][reset_idx][dest].append((reset_rect, hit_idx, reset_list[-1]))
+
+        possible_transitions = []
+        # Combine reset rects and construct transitions
+        for agent in reset_dict:
+            for reset_idx in reset_dict[agent]:
+                for dest in reset_dict[agent][reset_idx]:
+                    reset_data = tuple(map(list, zip(*reset_dict[agent][reset_idx][dest])))
+                    paths = [r[-1] for r in reset_data[-1]]
+                    transition = (agent, node.mode[agent],dest, *reset_data[:-1], paths)
+                    src_mode = node.get_mode(agent, node.mode[agent])
+                    src_track = node.get_track(agent, node.mode[agent])
+                    dest_mode = node.get_mode(agent, dest)
+                    dest_track = node.get_track(agent, dest)
+                    if dest_track == track_map.h(src_track, src_mode, dest_mode):
+                        possible_transitions.append(transition)
+                        print(transition[4])
+        # Return result
+        return None, possible_transitions
+
+    def apply_reset(self, agent: BaseAgent, reset_list, all_agent_state, track_map) -> Tuple[str, np.ndarray]:
+        dest = []
+        rect = []
+
+        agent_state, agent_mode, agent_static = all_agent_state[agent.id]
+
+        dest = copy.deepcopy(agent_mode)
+        possible_dest = [[elem] for elem in dest]
+        ego_type = find(agent.decision_logic.args, lambda a: a.name == EGO).typ
+        rect = copy.deepcopy([agent_state[0][1:], agent_state[1][1:]])
+
+        # The reset_list here are all the resets for a single transition. Need to evaluate each of them
+        # and then combine them together
+        for reset_tuple in reset_list:
+            reset, disc_var_dict, cont_var_dict, _, _p = reset_tuple
+            reset_variable = reset.var
+            expr = reset.expr
+            # First get the transition destinations
+            if "mode" in reset_variable:
+                found = False
+                for var_loc, discrete_variable_ego in enumerate(agent.decision_logic.state_defs[ego_type].disc):
+                    if discrete_variable_ego == reset_variable:
+                        found = True
+                        break
+                if not found:
+                    raise ValueError(
+                        f'Reset discrete variable {discrete_variable_ego} not found')
+                if isinstance(reset.val_ast, ast.Constant):
+                    val = eval(expr)
+                    possible_dest[var_loc] = [val]
+                else:
+                    tmp = expr.split('.')
+                    if 'map' in tmp[0]:
+                        for var in disc_var_dict:
+                            expr = expr.replace(var, f"'{disc_var_dict[var]}'")
+                        res = eval(expr)
+                        if not isinstance(res, list):
+                            res = [res]
+                        possible_dest[var_loc] = res
+                    else:
+                        expr = tmp
+                        if expr[0].strip(' ') in agent.decision_logic.mode_defs:
+                            possible_dest[var_loc] = [expr[1]]
+
+            # Assume linear function for continuous variables
+            else:
+                lhs = reset_variable
+                rhs = expr
+                found = False
+                for lhs_idx, cts_variable in enumerate(agent.decision_logic.state_defs[ego_type].cont):
+                    if cts_variable == lhs:
+                        found = True
+                        break
+                if not found:
+                    raise ValueError(
+                        f'Reset continuous variable {cts_variable} not found')
+                # substituting low variables
+
+                symbols = []
+                for var in cont_var_dict:
+                    if var in expr:
+                        symbols.append(var)
+
+                # TODO: Implement this function
+                # The input to this function is a list of used symbols and the cont_var_dict
+                # The ouput of this function is a list of tuple of values for each variable in the symbols list
+                # The function will explor all possible combinations of low bound and upper bound for the variables in the symbols list
+                comb_list = self._get_combinations(symbols, cont_var_dict)
+
+                lb = float('inf')
+                ub = -float('inf')
+
+                for comb in comb_list:
+                    val_dict = {}
+                    tmp = copy.deepcopy(expr)
+                    for symbol_idx, symbol in enumerate(symbols):
+                        tmp = tmp.replace(symbol, str(comb[symbol_idx]))
+                    res = eval(tmp, {}, val_dict)
+                    lb = min(lb, res)
+                    ub = max(ub, res)
+
+                rect[0][lhs_idx] = lb
+                rect[1][lhs_idx] = ub
+
+        all_dest = itertools.product(*possible_dest)
+        dest = []
+        for tmp in all_dest:
+            dest.append(tmp)
+
+        return dest, rect
+
+    def _get_combinations(self, symbols, cont_var_dict):
+        data_list = []
+        for symbol in symbols:
+            data_list.append(cont_var_dict[symbol])
+        comb_list = list(itertools.product(*data_list))
+        return comb_list
+
+    def get_transition_verify_opt(self, cache: Dict[str, CachedRTTrans], paths: PathDiffs, node: AnalysisTreeNode, track_map, sensor) -> Tuple[Optional[Dict[str, List[str]]], Optional[Dict[str, List[Tuple[str, List[str], List[float]]]]]]:
+
+        # For each agent
+        agent_guard_dict = defaultdict(list)
+        cached_guards = defaultdict(list)
+        min_trans_ind = None
+        cached_trans = defaultdict(list)
+        agent_dict = node.agent
+
+        if not cache:
+            paths = [(agent, p) for agent in node.agent.values() for p in agent.decision_logic.paths]
+        else:
+
+            # _transitions = [trans.transition for seg in cache.values() for trans in seg.transitions]
+            _transitions = [(aid, trans) for aid, seg in cache.items() for trans in seg.transitions if reach_trans_suit(trans.inits, node.init)]
+            # pp(("cached trans", len(_transitions)))
+            if len(_transitions) > 0:
+                min_trans_ind = min([t.transition for _, t in _transitions])
+                # TODO: check for asserts
+                cached_trans = [(aid, tran.mode, tran.dest, tran.reset, tran.reset_idx, tran.paths) for aid, tran in dedup(_transitions, lambda p: (p[0], p[1].mode, p[1].dest)) if tran.transition == min_trans_ind]
+                if len(paths) == 0:
+                    # print(red("full cache"))
+                    return None, cached_trans
+
+                path_transitions = defaultdict(int)
+                for seg in cache.values():
+                    for tran in seg.transitions:
+                        for p in tran.paths:
+                            path_transitions[p.cond] = max(path_transitions[p.cond], tran.transition)
+                for agent_id, segment in cache.items():
+                    agent = node.agent[agent_id]
+                    if len(agent.decision_logic.args) == 0:
+                        continue
+                    state_dict = {aid: (node.trace[aid][0], node.mode[aid], node.static[aid]) for aid in node.agent}
+
+                    agent_paths = dedup([p for tran in segment.transitions for p in tran.paths], lambda i: (i.var, i.cond, i.val))
+                    for path in agent_paths:
+                        cont_var_dict_template, discrete_variable_dict, length_dict = sensor.sense(
+                            agent, state_dict, track_map)
+                        reset = (path.var, path.val_veri)
+                        guard_expression = GuardExpressionAst([path.cond_veri])
+
+                        cont_var_updater = guard_expression.parse_any_all_new(
+                            cont_var_dict_template, discrete_variable_dict, length_dict)
+                        self.apply_cont_var_updater(
+                            cont_var_dict_template, cont_var_updater)
+                        guard_can_satisfied = guard_expression.evaluate_guard_disc(
+                            agent, discrete_variable_dict, cont_var_dict_template, track_map)
+                        if not guard_can_satisfied:
+                            continue
+                        cached_guards[agent_id].append((path, guard_expression, cont_var_updater, copy.deepcopy(discrete_variable_dict), reset, path_transitions[path.cond]))
+
+        # for aid, trace in node.trace.items():
+        #     if len(trace) < 2:
+        #         pp(("weird state", aid, trace))
+        for agent, path in paths:
+            if len(agent.decision_logic.args) == 0:
+                continue
+            agent_id = agent.id
+            state_dict = {aid: (node.trace[aid][0:2], node.mode[aid], node.static[aid]) for aid in node.agent}
+            cont_var_dict_template, discrete_variable_dict, length_dict = sensor.sense(
+                agent, state_dict, track_map)
+            # TODO-PARSER: Get equivalent for this function
+            # Construct the guard expression
+            guard_expression = GuardExpressionAst([path.cond_veri])
+
+            cont_var_updater = guard_expression.parse_any_all_new(
+                cont_var_dict_template, discrete_variable_dict, length_dict)
+            self.apply_cont_var_updater(
+                cont_var_dict_template, cont_var_updater)
+            guard_can_satisfied = guard_expression.evaluate_guard_disc(
+                agent, discrete_variable_dict, cont_var_dict_template, track_map)
+            if not guard_can_satisfied:
+                continue
+            agent_guard_dict[agent_id].append(
+                (guard_expression, cont_var_updater, copy.deepcopy(discrete_variable_dict), path))
+
+        trace_length = int(min(len(v) for v in node.trace.values()) // 2)
+        # pp(("trace len", trace_length, {a: len(t) for a, t in node.trace.items()}))
+        guard_hits = []
+        guard_hit = False
+        reduction_rate = 10
+        reduction_queue = [(0, trace_length, trace_length)]
+        # for idx, end_idx,combine_len in reduction_queue:
+        while reduction_queue:
+            idx, end_idx,combine_len = reduction_queue.pop()
+            reduction_needed = False
+            # print((idx, combine_len))
+            if min_trans_ind != None and idx >= min_trans_ind:
+                return None, cached_trans
+            any_contained = False
+            hits = []
+            # end_idx = min(idx+combine_len, trace_length)
+            state_dict = {aid: (combine_rect_v1(node.trace[aid][idx*2:end_idx*2]), node.mode[aid], node.static[aid]) for aid in node.agent}
+            
+            asserts = defaultdict(list)
+            for agent_id in agent_dict.keys():
+                agent: BaseAgent = agent_dict[agent_id]
+                if len(agent.decision_logic.args) == 0:
+                    continue
+                # if np.array(agent_state).ndim != 2:
+                #     pp(("weird state", agent_id, agent_state))
+                cont_vars, disc_vars, len_dict = sensor.sense(agent, state_dict, track_map)
+                resets = defaultdict(list)
+                # Check safety conditions
+                for i, a in enumerate(agent.decision_logic.asserts_veri):
+                    pre_expr = a.pre
+
+                    def eval_expr(expr):
+                        ge = GuardExpressionAst([copy.deepcopy(expr)])
+                        cont_var_updater = ge.parse_any_all_new(cont_vars, disc_vars, len_dict)
+                        self.apply_cont_var_updater(cont_vars, cont_var_updater)
+                        sat = ge.evaluate_guard_disc(agent, disc_vars, cont_vars, track_map)
+                        if sat:
+                            sat = ge.evaluate_guard_hybrid(agent, disc_vars, cont_vars, track_map)
+                            if sat:
+                                sat, contained = ge.evaluate_guard_cont(agent, cont_vars, track_map)
+                                sat = sat and contained
+                        return sat
+                    if eval_expr(pre_expr):
+                        if not eval_expr(a.cond):
+                            if combine_len == 1:
+                                label = a.label if a.label != None else f"<assert {i}>"
+                                print(f"assert hit for {agent_id}: \"{label}\"")
+                                print(idx)
+                                asserts[agent_id].append(label)
+                            else:
+                                new_len = int(np.ceil(combine_len/reduction_rate))
+                                next_list = [(i, min(i+new_len, end_idx) ,new_len) for i in range(idx,end_idx,new_len)]
+                                reduction_queue.extend(next_list[::-1])
+                                reduction_needed = True
+                                break
+                if reduction_needed:
+                    break
+                if agent_id in asserts:
+                    continue
+                if agent_id not in agent_guard_dict:
+                    continue
+
+                unchecked_cache_guards = [g[:-1] for g in cached_guards[agent_id] if g[-1] < idx]     # FIXME: off by 1?
+                for guard_expression, continuous_variable_updater, discrete_variable_dict, path in agent_guard_dict[agent_id] + unchecked_cache_guards:
+                    assert isinstance(path, ModePath)
+                    new_cont_var_dict = copy.deepcopy(cont_vars)
+                    one_step_guard: GuardExpressionAst = copy.deepcopy(guard_expression)
+
+                    self.apply_cont_var_updater(new_cont_var_dict, continuous_variable_updater)
+                    guard_can_satisfied = one_step_guard.evaluate_guard_hybrid(
+                        agent, discrete_variable_dict, new_cont_var_dict, track_map)
+                    if not guard_can_satisfied:
+                        continue
+                    guard_satisfied, is_contained = one_step_guard.evaluate_guard_cont(
+                        agent, new_cont_var_dict, track_map)
+                    if combine_len == 1:
+                        any_contained = any_contained or is_contained
+                    # TODO: Can we also store the cont and disc var dict so we don't have to call sensor again?
+                    if guard_satisfied and combine_len == 1:
+                        reset_expr = ResetExpression((path.var, path.val_veri))
+                        resets[reset_expr.var].append(
+                            (reset_expr, discrete_variable_dict,
+                             new_cont_var_dict, guard_expression.guard_idx, path)
+                        )
+                    elif guard_satisfied and combine_len > 1:       
+                        new_len = int(np.ceil(combine_len/reduction_rate))
+                        next_list = [(i, min(i+new_len, end_idx) ,new_len) for i in range(idx,end_idx,new_len)]
+                        reduction_queue.extend(next_list[::-1])
+                        reduction_needed = True
+                        break
+                if reduction_needed:
+                    break
+                if combine_len == 1:
+                    # Perform combination over all possible resets to generate all possible real resets
+                    combined_reset_list = list(itertools.product(*resets.values()))
+                    if len(combined_reset_list) == 1 and combined_reset_list[0] == ():
+                        continue
+                    for i in range(len(combined_reset_list)):
+                        # Compute reset_idx
+                        reset_idx = []
+                        for reset_info in combined_reset_list[i]:
+                            reset_idx.append(reset_info[3])
+                        # a list of reset expression
+                        hits.append((agent_id, tuple(reset_idx), combined_reset_list[i]))
+            
+            if reduction_needed or combine_len > 1:
+                continue
+            if len(asserts) > 0:
+                return (asserts, idx), None
+            if hits != []:
+                guard_hits.append((hits, state_dict, idx))
+                guard_hit = True
+            elif guard_hit:
+                break
+            if any_contained:
+                break
+
+        reset_dict = {}  # defaultdict(lambda: defaultdict(list))
+        for hits, all_agent_state, hit_idx in guard_hits:
+            for agent_id, reset_idx, reset_list in hits:
+                # TODO: Need to change this function to handle the new reset expression and then I am done
+                dest_list, reset_rect = self.apply_reset(node.agent[agent_id], reset_list, all_agent_state, track_map)
+                # pp(("dests", dest_list, *[astunparser.unparse(reset[-1].val_veri) for reset in reset_list]))
+                if agent_id not in reset_dict:
+                    reset_dict[agent_id] = {}
+                if not dest_list:
+                    warnings.warn(
+                        f"Guard hit for mode {node.mode[agent_id]} for agent {agent_id} without available next mode")
+                    dest_list.append(None)
+                if reset_idx not in reset_dict[agent_id]:
+                    reset_dict[agent_id][reset_idx] = {}
+                for dest in dest_list:
+                    if dest not in reset_dict[agent_id][reset_idx]:
+                        reset_dict[agent_id][reset_idx][dest] = []
+                    reset_dict[agent_id][reset_idx][dest].append((reset_rect, hit_idx, reset_list[-1]))
+
+        possible_transitions = []
+        # Combine reset rects and construct transitions
+        for agent in reset_dict:
+            for reset_idx in reset_dict[agent]:
+                for dest in reset_dict[agent][reset_idx]:
+                    reset_data = tuple(map(list, zip(*reset_dict[agent][reset_idx][dest])))
+                    paths = [r[-1] for r in reset_data[-1]]
+                    transition = (agent, node.mode[agent],dest, *reset_data[:-1], paths)
+                    src_mode = node.get_mode(agent, node.mode[agent])
+                    src_track = node.get_track(agent, node.mode[agent])
+                    dest_mode = node.get_mode(agent, dest)
+                    dest_track = node.get_track(agent, dest)
+                    if dest_track == track_map.h(src_track, src_mode, dest_mode):
+                        possible_transitions.append(transition)
+                        # print(transition[4])
+        # Return result
+        return None, possible_transitions
 
 def checkHeight(root, max_height):
     if root:
